View Bookings page for tour operators
These changes ensure that the View Bookings page for the tour operators
in now in sync with the new look of the page in the base plugin.
Have also removed some commented code in the plugin and fixed a few
warnings that were visible in the tour operators tab in the admin
product page.