Compatible with Wootheme booking plugin.
Issue with WooThemes booking plugin and our booking plugin, if both
plugins are active then user can not able to add the product to the cart
which have our booking plugin setting.

Now the booking plugin and addon is compatible with the wootheme booking
plugin. With this fix it Allow you to activate both “WooTheme booking”
plugin and our booking plugin together on site.

The fix will work with the product which have either our booking plugin
and addon setting enabled or WooThemes booking plugin setting enable.
This fix will allow to add the product to cart which have our booking
plugin and addon setting enabled.

If same product have both plugins settings enabled then it will not
work. So one have to keep both plugin setting on different products.