Addon compatibility with GCal 2 way Sync
The below changes have been included to allow 2 way sync with the google
calendar.

The below changes allow for 2 way sync with the google calendar for only
exporting. Importing changes shall be uploaded later.

The tour operator can setup sync settings only if it is allowed by the
admin. Also, the key file for the tour operator should be uploaded in
the key folder for the base plugin i.e.
/woocommerce-booking/includes/gcal/key/