Import Bookings for GCal 2 way sync
The tour operator can now import bookings manually using the Import
Events button on his/her profile page.

Alternately, the events from the Google Calendar saved in the Profile
page will be automatically imported once every 24 hours.

All the Imported Events will be listed in the Booking->Import Bookings
page.

They need to be mapped to the correct product listed in the drop down on
the page.

It should be noted that the Tour Operator shall be able to see only the
products assigned to him in the drop down.