Fix #13
I have added a global variable in the print tickets class. This variable
is set to Yes, when the email is being sent to the tour operator. After
the email is sent it's value is updated to No.

So, before adding the buttons, we always check the variable value and
add the buttons based on the same.