Import Bookings page Product Dropdown
The Product Drop down on the Import bookings page displayed only the
base product for variable products. The variations were not displayed as
individual products to help select a given variation when mapping.

This has now been fixed to ensure that each variation is displayed in
the dropdown individually to enable ease of mapping.