Version 1.7
Change Log Tour Operators Addon v1.7

This update to the plugin contains new enhancements and compatibility
changes with the Booking & Appointment Plugin for WooCommerce.

Enhancements:

1. The addon has now been made compatible with the Booking confirmations
feature. So, now the tour operators will receive an email whenever a
booking for a product (which is assigned to them) is received and the
order needs confirmation.

2. The Tour Operator shall have the ability to approve or reject the
booking.

3. The addon has also been made compatible to allow 2 way Google
Calendar Sync for the Tour Operators.

4. The Tour Operators can now setup their Google Calendar sync settings
from their Profile page.