Fix #24
The issue arises as when paginating the records in the base plugin,
we're only sending chunks of 30 records to the addon.

The addon also creates the number of pages based on the total number of
records for the tour operator.

So for e.g. if the admin has 1000 records of which 49 are for this tour
operator.

So the tour operator can see only 2 pages, so he's basically able to see
his records from only the first 60 records.

For now, I've removed the code where we're setting the number of pages
for the tour operator.

So basically, the operator will be able to see his 49 records spread
over as many pages as the admin, i.e. he'll be able to see 3 records on
the first page, none on the second page, 4 on the third page and so on.