Import Bookings for Tour Operator
In this commit, I have added code to create the post of the events which
are imported by the the Tour Operator.

So now onwards, all the events which are imported will be saved in the
post.

Also, the Tour Operator can only see their events on the Import Bookings
page and map with their products only.

Removed the submenu from the addon file as now we are using post view to
display the events and bookings to tour opertor so no need to again add
the submenu in this addon.

Removed old Import Bookings page as now there is no need to keep that
code because all the events are now stored as post and we are using post
view to show those events hence removed the old code.

Also, added a fix for the issue elaborated as below.
If the ics feed url set to import at tour level is start with https then
no events were being imported. The same has been fixed in this commit.